语言热度榜及雇主需求榜部分代码

+++ src/main/java/com/wizzstudio/languagerank/dto/FinalRankDTO.java
@@ -10,7 +10,7 @@ public class FinalRankDTO {
     private Language language;
 
     // 语言热度指数
-    private Integer FinalExponent;
+    private Double FinalExponent;
 
     // 语言热度变化趋势(值在LanguageTendEnum类中列出)
     private Integer languageTend;
